create new separate class for UIControl extension
