+++ FTPopOverMenu_Swift/UIControl+Point.swift
@@ -1,0 +1,33 @@
+//
+//  UIControl+Point.swift
+//  FTPopOverMenu_Swift
+//
+//  Created by Abdullah Selek on 28/07/2017.
+//  Copyright © 2017 LiuFengting. All rights reserved.
+//
+
+import UIKit
+
+extension UIControl {
+
+    // solution found at: http://stackoverflow.com/a/5666430/6310268
+
+    internal func setAnchorPoint(anchorPoint: CGPoint) {
+        var newPoint = CGPoint(x: self.bounds.size.width * anchorPoint.x, y: self.bounds.size.height * anchorPoint.y)
+        var oldPoint = CGPoint(x: self.bounds.size.width * self.layer.anchorPoint.x, y: self.bounds.size.height * self.layer.anchorPoint.y)
+
+        newPoint = newPoint.applying(self.transform)
+        oldPoint = oldPoint.applying(self.transform)
+
+        var position = self.layer.position
+        position.x -= oldPoint.x
+        position.x += newPoint.x
+
+        position.y -= oldPoint.y
+        position.y += newPoint.y
+
+        self.layer.position = position
+        self.layer.anchorPoint = anchorPoint
+    }
+
+}
